Get Activity

Returns the given activity that is owned by the authenticated athlete. Requires activity:read for Everyone and Followers activities. Requires activity:read_all for Only Me activities.

Usage

get_activities(
  id,
  include_all_efforts = NULL,
  scope = "activity:read",
  client = strava_client()
)

Arguments

id

The identifier of the activity.

include_all_efforts

To include all segments efforts.

scope

Authorization scope.

client

Strava client. By default, assumes you have client credentials stored in your .Renviron file.

Value

The activity's detailed representation. An instance of DetailedActivity.

Examples

if (FALSE) {
  get_activities(7255352296)
}

tidy(strava_data$activities)
#> # A tibble: 1 × 68
#>   resource_state athlete   name         distance moving_time elapsed_time
#>            <int> <list>    <chr>           <dbl>       <int>        <int>
#> 1              3 <int [1]> Morning Ride   56930.        8218         9489
#> # … with 62 more variables: total_elevation_gain <dbl>, type <chr>,
#> #   sport_type <chr>, workout_type <int>, id <dbl>, start_date <chr>,
#> #   start_date_local <chr>, timezone <chr>, utc_offset <dbl>,
#> #   location_city <lgl>, location_state <lgl>, location_country <chr>,
#> #   achievement_count <int>, kudos_count <int>, comment_count <int>,
#> #   athlete_count <int>, photo_count <int>, map <list>, trainer <lgl>,
#> #   commute <lgl>, manual <lgl>, private <lgl>, visibility <chr>, …