Create a strava client

Strava uses OAuth2 for authentication to the V3 API. OAuth allows external applications to request authorization to a user's data. It allows users to grant and revoke API access on a per-application basis and keeps users' authentication details safe. All developers need to register their application before getting started. A registered application will be assigned a client ID and client secret. The secret is used for authentication and should never be shared.

Usage

strava_client(
  id = Sys.getenv("STRAVA_ID"),
  secret = Sys.getenv("STRAVA_SECRET"),
  token_url = "https://www.strava.com/oauth/token",
  name = "strava-oauth",
  ...
)

Arguments

id

Client ID.

secret

Client secret.

token_url

Token URL.

name

Client name.

...

Additional parameters passed to httr2::oauth_client.

See also

Examples

# strava client will pull environment variables by default
Sys.setenv(
  STRAVA_ID="123",
  STRAVA_SECRET="my_secret_code"
)

strava_client()
#> <httr2_oauth_client>
#> name: strava-oauth
#> id: 123
#> secret: <REDACTED>
#> token_url: https://www.strava.com/oauth/token
#> auth: oauth_client_req_auth_body